refactor: Modularization ver.1
